Add delegated<Event>Bubble property
For every type of event an additional property can be used, denoted by
`delegated<Event>Bubble`, which can be set to true to explicitly allow
bubbling in scenarios where it is needed. This relates to a previous bug
detailed here:
https://github.com/rniemeyer/knockout-delegatedEvents/issues/7

A common use case of this is for example on Bootstrap dropdown menus,
which require the event to bubble in order for the dropdown to properly
capture the click event and close as expected.